feat: add functions to group app options by projects and environments by clusters
