+++ src/Common/Common.service.ts
@@ -23,7 +23,7 @@ import {
     sanitizeUserApprovalList,
     stringComparatorBySortOrder,
 } from '@Shared/Helpers'
-import { PolicyBlockInfo, RuntimeParamsAPIResponseType, RuntimePluginVariables } from '@Shared/types'
+import { EnvListMinDTO, PolicyBlockInfo, RuntimeParamsAPIResponseType, RuntimePluginVariables } from '@Shared/types'
 import { GitProviderType, ROUTES } from './Constants'
 import { getUrlWithSearchParams, sortCallback } from './Helper'
 import {
@@ -51,11 +51,14 @@ import {
     GetAppsInfoForEnvProps,
     AppMeta,
     ApprovalRuntimeStateType,
+    EnvironmentsGroupedByClustersType,
+    AppsGroupedByProjectsType,
 } from './Types'
 import { ApiResourceType, STAGE_MAP } from '../Pages'
 import { RefVariableType, VariableTypeFormat } from './CIPipeline.Types'
 import { get, post } from './API'
 import { StatusType } from '@Shared/Components'
+import { EnvironmentTypeEnum } from '@Shared/constants'
 
 export const getTeamListMin = (): Promise<TeamList> => {
     // ignore active field
@@ -563,4 +566,68 @@ export const getAppsInfoForEnv = async ({ envId, appIds }: GetAppsInfoForEnvProp
             return agg
         }, []),
     }
+}
+
+
+export const getAppOptionsGroupedByProjects = async (): Promise<AppsGroupedByProjectsType> => {
+    const { result } = await get<AppsGroupedByProjectsType>(ROUTES.APP_LIST_MIN)
+
+    if (!result) {
+        return []
+    }
+
+    return result
+        .map((project) => ({
+            ...project,
+            appList: project.appList.sort((a, b) => stringComparatorBySortOrder(a.name, b.name)),
+        }))
+        .sort((a, b) => stringComparatorBySortOrder(a.projectName, b.projectName))
+}
+
+
+export const getEnvironmentOptionsGroupedByClusters = async (): Promise<EnvironmentsGroupedByClustersType> => {
+    const { result } = (await get(ROUTES.ENVIRONMENT_LIST_MIN)) as ResponseType<EnvListMinDTO[]>
+
+    if (!result) {
+        return []
+    }
+
+    const sortedEnvList = result
+        .map(
+            ({
+                id,
+                environment_name: name,
+                isVirtualEnvironment,
+                cluster_name: cluster,
+                default: isDefault,
+                namespace,
+            }) => ({
+                id,
+                name,
+                isVirtual: isVirtualEnvironment ?? false,
+                cluster,
+                environmentType: isDefault ? EnvironmentTypeEnum.production : EnvironmentTypeEnum.nonProduction,
+                namespace,
+            }),
+        )
+        .sort((a, b) => stringComparatorBySortOrder(a.name, b.name))
+
+    const envGroupedByCluster = Object.values(
+        sortedEnvList.reduce<
+            Record<EnvironmentsGroupedByClustersType[number]['clusterName'], EnvironmentsGroupedByClustersType[number]>
+        >((acc, env) => {
+            if (!acc[env.cluster]) {
+                acc[env.cluster] = {
+                    clusterName: env.cluster,
+                    envList: [],
+                }
+            }
+
+            acc[env.cluster].envList.push(env)
+
+            return acc
+        }, {}),
+    ).sort((a, b) => stringComparatorBySortOrder(a.clusterName, b.clusterName))
+
+    return envGroupedByCluster
 }
