+++ src/Shared/Hooks/useGetResourceKindsOptions/service.ts
@@ -14,30 +14,9 @@
  * limitations under the License.
  */
 
-import { get } from '@Common/API'
 import { getTeamListMin } from '@Common/Common.service'
-import { ROUTES } from '@Common/Constants'
-import { ResponseType, Teams } from '@Common/Types'
-import { EnvironmentTypeEnum } from '@Shared/constants'
+import { Teams } from '@Common/Types'
 import { stringComparatorBySortOrder } from '@Shared/Helpers'
-import { EnvListMinDTO } from '@Shared/types'
-
-import { AppsGroupedByProjectsType, EnvironmentsGroupedByClustersType } from './types'
-
-export const getAppOptionsGroupedByProjects = async (): Promise<AppsGroupedByProjectsType> => {
-    const { result } = (await get(ROUTES.APP_LIST_MIN)) as ResponseType<AppsGroupedByProjectsType>
-
-    if (!result) {
-        return []
-    }
-
-    return result
-        .map((project) => ({
-            ...project,
-            appList: project.appList.sort((a, b) => stringComparatorBySortOrder(a.name, b.name)),
-        }))
-        .sort((a, b) => stringComparatorBySortOrder(a.projectName, b.projectName))
-}
 
 export const getProjectOptions = async (): Promise<Pick<Teams, 'id' | 'name'>[]> => {
     const { result } = await getTeamListMin()
@@ -52,51 +31,4 @@ export const getProjectOptions = async (): Promise<Pick<Teams, 'id' | 'name'>[]>
             name,
         }))
         .sort((a, b) => stringComparatorBySortOrder(a.name, b.name))
-}
-
-export const getEnvironmentOptionsGroupedByClusters = async (): Promise<EnvironmentsGroupedByClustersType> => {
-    const { result } = (await get(ROUTES.ENVIRONMENT_LIST_MIN)) as ResponseType<EnvListMinDTO[]>
-
-    if (!result) {
-        return []
-    }
-
-    const sortedEnvList = result
-        .map(
-            ({
-                id,
-                environment_name: name,
-                isVirtualEnvironment,
-                cluster_name: cluster,
-                default: isDefault,
-                namespace,
-            }) => ({
-                id,
-                name,
-                isVirtual: isVirtualEnvironment ?? false,
-                cluster,
-                environmentType: isDefault ? EnvironmentTypeEnum.production : EnvironmentTypeEnum.nonProduction,
-                namespace,
-            }),
-        )
-        .sort((a, b) => stringComparatorBySortOrder(a.name, b.name))
-
-    const envGroupedByCluster = Object.values(
-        sortedEnvList.reduce<
-            Record<EnvironmentsGroupedByClustersType[number]['clusterName'], EnvironmentsGroupedByClustersType[number]>
-        >((acc, env) => {
-            if (!acc[env.cluster]) {
-                acc[env.cluster] = {
-                    clusterName: env.cluster,
-                    envList: [],
-                }
-            }
-
-            acc[env.cluster].envList.push(env)
-
-            return acc
-        }, {}),
-    ).sort((a, b) => stringComparatorBySortOrder(a.clusterName, b.clusterName))
-
-    return envGroupedByCluster
 }
